requirements.txt
tensorboard==2.11.2
accelerate==0.20.3
transformers==4.27.4
hydra-core==1.3.0
einops==0.3.2
pandas==2.0.0
open3d==0.19.0
Pillow==11.1.0
torchpack==0.3.1
nuscenes-devkit==1.1.10
mpi4py==3.1.4
scipy==1.10.1
numpy==1.21.6
numba==0.53.0
llvmlite==0.36.0
opencv-python==4.6.0.66
moviepy==1.0.3
huggingface-hub==0.24.6
xformers==0.0.19

sentencepiece
blobfile
matplotlib
prettytable
tensorboardX
scikit-learn
tqdm
pyyaml
timm